+++ app/serializers/export/own_post_serializer.rb
@@ -29,7 +29,7 @@ module Export
     end
 
     def excluded_subscription_key
-      entity.public ? :subscribed_users_ids : :subscribed_pods_uris
+      object.public? ? :subscribed_users_ids : :subscribed_pods_uris
     end
   end
 end