+++ spec/support/serializer_matchers.rb
@@ -15,6 +15,7 @@
 RSpec::Matchers.define :serialize_association do |association_name|
   match do |root_serializer_class|
     association = fetch_association(root_serializer_class, association_name)
+    @serializer_from_options = association.serializer_from_options
     execute_receive_matcher_with(association)
   end
 
@@ -50,7 +51,11 @@ RSpec::Matchers.define :serialize_association do |association_name|
 
   def with_object_expectation(object)
     if association_object.is_a?(Array)
-      expect(object).to match_array(association_object)
+      if serializer_class == FlatMapArraySerializer
+        expect(object.flat_map(&:to_a)).to match_array(association_object)
+      else
+        expect(object).to match_array(association_object)
+      end
     elsif !association_object.nil?
       expect(object).to eq(association_object)
     end
@@ -66,6 +71,7 @@ RSpec::Matchers.define :serialize_association do |association_name|
 
   def pick_serializer_class
     return association_serializer_class unless association_serializer_class.nil?
+    return @serializer_from_options unless @serializer_from_options.nil?
     return ActiveModel::ArraySerializer unless each_serializer_class.nil?
   end
 end