+++ app/presenters/node_info_presenter.rb
@@ -24,6 +24,7 @@ class NodeInfoPresenter
     doc.open_registrations       = open_registrations?
     doc.metadata["nodeName"]     = name
     doc.metadata["xmppChat"]     = chat_enabled?
+    doc.metadata["camo"]         = camo_config
     doc.metadata["adminAccount"] = admin_account
   end
 
@@ -71,6 +72,14 @@ class NodeInfoPresenter
 
   def chat_enabled?
     AppConfig.chat.enabled?
+  end
+
+  def camo_config
+    {
+      markdown:   AppConfig.privacy.camo.proxy_markdown_images?,
+      opengraph:  AppConfig.privacy.camo.proxy_opengraph_thumbnails?,
+      remotePods: AppConfig.privacy.camo.proxy_remote_pod_images?
+    }
   end
 
   def admin_account