+++ app/mailers/notification_mailers/started_sharing.rb
@@ -2,7 +2,7 @@
 
 module NotificationMailers
   class StartedSharing < NotificationMailers::Base
-    def set_headers(*_args) # rubocop:disable Style/AccessorMethodName
+    def set_headers(*_args) # rubocop:disable Naming/AccessorMethodName
       @headers[:subject] = I18n.t("notifier.started_sharing.subject", name: @sender.name)
     end
   end